Major code refactor! Laying foundation for Chat bot.
