+++ internal/config/consts.go
@@ -1,0 +1,33 @@
+package config
+
+const (
+	ReplaceHighlightedText     = "ReplaceHighlightedText"
+	SpeakAIResponseKey         = "SpeakAIResponseKey"
+	ShowPopUpKey               = "ShowPopUpKey"
+	ShowStartWindowKey         = "showStartWindow"
+	firstRunKey                = "firstRun"
+	CurrentPromptKey           = "lastPrompt"
+	CurrentModelKey            = "lastModel"
+	CurrentFromLangKey         = "fromLang"
+	CurrentToLangKey           = "toLang"
+	StopOllamaOnShutDownKey    = "stopOllamaOnShutDown"
+	UseRemoteOllamaKey         = "useRemoteOllama"
+	OllamaURLKey               = "OllamaURL"
+	UseDockerKey               = "useDocker"
+	AskAIKeyboardShortcut      = "AskAIKeyboardShortcut"
+	CtrlReviseKeyboardShortcut = "CtrlReviseKeyboardShortcut"
+	TranslateKeyboardShortcut  = "TranslateKeyboardShortcut"
+
+	ConsumersKey = "ConsumersCook"
+	MealKey      = "MealToCook"
+	FlavorKey    = "MealFlavor"
+	CookwareKey  = "MealCookware"
+	DairyKey     = "MealDairy"
+	FreezerKey   = "MealFreezer"
+	HerbsKey     = "MealHerbs"
+	PantryKey    = "MealPantry"
+	VeggiesKey   = "MealVeggies"
+	ProteinKey   = "MealProtein"
+
+	LengthOfKeyBoardShortcuts = 3
+)
