+++ internal/gui/food/meal_planner.go
@@ -1,9 +1,12 @@
-package main
+package food
 
 import (
 	"crypto/sha256"
+	"github.com/bahelit/ctrl_plus_revise/internal/gui/loading"
+	"github.com/bahelit/ctrl_plus_revise/internal/gui/menu"
 	"log/slog"
 	"math/rand"
+	"net/url"
 	"sort"
 	"time"
 
@@ -14,7 +17,9 @@ import (
 	"fyne.io/fyne/v2/layout"
 	"fyne.io/fyne/v2/theme"
 	"fyne.io/fyne/v2/widget"
-	"github.com/bahelit/ctrl_plus_revise/internal/gui"
+
+	"github.com/bahelit/ctrl_plus_revise/internal/config"
+	"github.com/bahelit/ctrl_plus_revise/internal/gui/shortcuts"
 	"github.com/bahelit/ctrl_plus_revise/internal/ollama"
 	ollamaApi "github.com/ollama/ollama/api"
 )
@@ -33,7 +38,7 @@ type MealInfo struct {
 	Allergies []string
 }
 
-func mealPlanner(guiApp fyne.App) {
+func MealPlanner(guiApp fyne.App, ollamaClient *ollamaApi.Client) {
 	var (
 		screenHeight float32 = 575.0
 		screenWidth  float32 = 655.0
@@ -41,7 +46,7 @@ func mealPlanner(guiApp fyne.App) {
 	mealPlanner := guiApp.NewWindow("Ctrl+Revise Meal Planner")
 	mealPlanner.Resize(fyne.NewSize(screenWidth, screenHeight))
 
-	mealPlanner.SetMainMenu(makeMenu(guiApp, mealPlanner))
+	mealPlanner.SetMainMenu(menu.MakeMenu(guiApp, ollamaClient, mealPlanner))
 
 	var mealInfo MealInfo
 
@@ -50,20 +55,20 @@ func mealPlanner(guiApp fyne.App) {
 
 	meal := widget.NewRadioGroup(getMeals(), func(s string) {
 		slog.Debug("selected meal", "meal", s)
-		guiApp.Preferences().SetString(MealKey, s)
+		guiApp.Preferences().SetString(config.MealKey, s)
 	})
-	meal.SetSelected(guiApp.Preferences().String(MealKey))
+	meal.SetSelected(guiApp.Preferences().String(config.MealKey))
 	meal.OnChanged = func(s string) {
-		guiApp.Preferences().SetString(MealKey, s)
+		guiApp.Preferences().SetString(config.MealKey, s)
 		mealInfo.Meal = s
 	}
 
 	consumersValidated := newNumEntry()
 	consumersValidated.SetPlaceHolder("How many people are eating?")
 	consumersValidated.OnSubmitted = func(s string) {
-		guiApp.Preferences().SetString(ConsumersKey, s)
+		guiApp.Preferences().SetString(config.ConsumersKey, s)
 	}
-	consumers := guiApp.Preferences().String(ConsumersKey)
+	consumers := guiApp.Preferences().String(config.ConsumersKey)
 	if consumers != "" {
 		consumersValidated.SetText(consumers)
 	}
@@ -73,62 +78,62 @@ func mealPlanner(guiApp fyne.App) {
 
 	flavor := widget.NewRadioGroup(getThemes(), func(s string) {
 		slog.Debug("selected cuisine", "flavor", s)
-		guiApp.Preferences().SetString(FlavorKey, s)
+		guiApp.Preferences().SetString(config.FlavorKey, s)
 	})
-	flavor.SetSelected(guiApp.Preferences().String(FlavorKey))
+	flavor.SetSelected(guiApp.Preferences().String(config.FlavorKey))
 	flavor.OnChanged = func(s string) {
-		guiApp.Preferences().SetString(FlavorKey, s)
+		guiApp.Preferences().SetString(config.FlavorKey, s)
 		mealInfo.Theme = s
 	}
 	flavorCard := widget.NewCard("Flavor Profile", "What kind of style for our food?", flavor)
 
 	cookWare := widget.NewCheckGroup(getCookWares(), func(strings []string) {
 		slog.Debug("selected cookWare", "cookWare", strings)
-		guiApp.Preferences().SetStringList(CookwareKey, strings)
+		guiApp.Preferences().SetStringList(config.CookwareKey, strings)
 	})
-	cookWare.SetSelected(guiApp.Preferences().StringList(CookwareKey))
+	cookWare.SetSelected(guiApp.Preferences().StringList(config.CookwareKey))
 	cookwareCard := widget.NewCard("Cookware", "What do we have cooking with?", cookWare)
 
 	dairyCheck := widget.NewCheckGroup(getDairy(), func(strings []string) {
 		slog.Debug("dairy stuff", "dairy", strings)
-		guiApp.Preferences().SetStringList(DairyKey, strings)
+		guiApp.Preferences().SetStringList(config.DairyKey, strings)
 	})
-	dairyCheck.SetSelected(guiApp.Preferences().StringList(DairyKey))
+	dairyCheck.SetSelected(guiApp.Preferences().StringList(config.DairyKey))
 	dairyCard := widget.NewCard("Dairy", "Cool the burn with some dairy", dairyCheck)
 
 	freezerCheck := widget.NewCheckGroup(getFreezerStuffs(), func(strings []string) {
 		slog.Debug("stuff in freezer", "stuff in freezer", strings)
-		guiApp.Preferences().SetStringList(FreezerKey, strings)
+		guiApp.Preferences().SetStringList(config.FreezerKey, strings)
 	})
-	freezerCheck.SetSelected(guiApp.Preferences().StringList(FreezerKey))
+	freezerCheck.SetSelected(guiApp.Preferences().StringList(config.FreezerKey))
 	freezerCard := widget.NewCard("freezer", "What's in the freezer?", freezerCheck)
 
 	herbsCheck := widget.NewCheckGroup(getHerbsSpices(), func(strings []string) {
 		slog.Debug("herbs & spices", "herbs", strings)
-		guiApp.Preferences().SetStringList(HerbsKey, strings)
+		guiApp.Preferences().SetStringList(config.HerbsKey, strings)
 	})
-	herbsCheck.SetSelected(guiApp.Preferences().StringList(HerbsKey))
+	herbsCheck.SetSelected(guiApp.Preferences().StringList(config.HerbsKey))
 	herbsCard := widget.NewCard("Pantry", "Enhance the aroma and flavor.", herbsCheck)
 
 	pantryCheck := widget.NewCheckGroup(getPantryStuffs(), func(strings []string) {
 		slog.Debug("stuff in pantry", "pantry", strings)
-		guiApp.Preferences().SetStringList(PantryKey, strings)
+		guiApp.Preferences().SetStringList(config.PantryKey, strings)
 	})
-	pantryCheck.SetSelected(guiApp.Preferences().StringList(PantryKey))
+	pantryCheck.SetSelected(guiApp.Preferences().StringList(config.PantryKey))
 	pantryCard := widget.NewCard("Pantry", "What's in the pantry?", pantryCheck)
 
 	veggieCheck := widget.NewCheckGroup(getVegetables(), func(strings []string) {
 		slog.Debug("dont' forget your vegetables", "pantry", strings)
-		guiApp.Preferences().SetStringList(VeggiesKey, strings)
+		guiApp.Preferences().SetStringList(config.VeggiesKey, strings)
 	})
-	veggieCheck.SetSelected(guiApp.Preferences().StringList(VeggiesKey))
+	veggieCheck.SetSelected(guiApp.Preferences().StringList(config.VeggiesKey))
 	veggieCard := widget.NewCard("Vegetables", "Don't forget you're vegetables.", veggieCheck)
 
 	proteinCheck := widget.NewCheckGroup(getProteins(), func(strings []string) {
 		slog.Debug("Meat or meat alternatives", "pantry", strings)
-		guiApp.Preferences().SetStringList(ProteinKey, strings)
+		guiApp.Preferences().SetStringList(config.ProteinKey, strings)
 	})
-	proteinCheck.SetSelected(guiApp.Preferences().StringList(ProteinKey))
+	proteinCheck.SetSelected(guiApp.Preferences().StringList(config.ProteinKey))
 	proteinCard := widget.NewCard("Protein", "Meat or meat substitutes", proteinCheck)
 
 	allergyCheck := widget.NewCheckGroup(getAllergies(), func(strings []string) {
@@ -162,81 +167,77 @@ func mealPlanner(guiApp fyne.App) {
 	suggest := widget.NewButton("Suggest a Single Meal", func() {
 		recipe := createMealPrompt(mealInfo)
 		slog.Info("Recipe for single meal", "PROMPT", recipe)
-		model := guiApp.Preferences().IntWithFallback(CurrentModelKey, int(ollama.Llama3Dot1))
-		loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-			"Preparing recipe with model: "+ollama.ModelName(model).String())
+		loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+			"Preparing recipe")
 		loadingScreen.Show()
 
-		recipe = addMarkdownFomatingToRecipe(recipe)
+		recipe = addMarkdownFormattingToRecipe(recipe)
 
-		generated, err := ollama.AskAI(ollamaClient, ollama.ModelName(model), recipe)
+		generated, err := ollama.AskAI(guiApp, ollamaClient, recipe)
 		if err != nil {
 			slog.Error("Failed to ask AI", "error", err)
 			loadingScreen.Hide()
 			return
 		}
 		loadingScreen.Hide()
-		recipePopUp(guiApp, recipe, &generated)
+		recipePopUp(guiApp, ollamaClient, recipe, &generated)
 	})
 	suggest.Importance = widget.HighImportance
 	suggestPrep := widget.NewButton("Prep Multiple Meals", func() {
 		recipe := createMealPrepPrompt(mealInfo)
 		slog.Info("Meal prep plan", "PROMPT", recipe)
-		model := guiApp.Preferences().IntWithFallback(CurrentModelKey, int(ollama.Llama3Dot1))
-		loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-			"Creating meal prep with model: "+ollama.ModelName(model).String())
+		loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+			"Creating meal prep")
 		loadingScreen.Show()
 
-		recipe = addMarkdownFomatingToRecipe(recipe)
+		recipe = addMarkdownFormattingToRecipe(recipe)
 
-		generated, err := ollama.AskAI(ollamaClient, ollama.ModelName(model), recipe)
+		generated, err := ollama.AskAI(guiApp, ollamaClient, recipe)
 		if err != nil {
 			slog.Error("Failed to ask AI", "error", err)
 			loadingScreen.Hide()
 			return
 		}
 		loadingScreen.Hide()
-		recipePopUp(guiApp, recipe, &generated)
+		recipePopUp(guiApp, ollamaClient, recipe, &generated)
 	})
 	suggestPrep.Importance = widget.HighImportance
 	groceryList := widget.NewButton("Create a Grocery List", func() {
 		recipe := createGroceryListPrompt(mealInfo)
 		slog.Info("Grocery List", "PROMPT", recipe)
-		model := guiApp.Preferences().IntWithFallback(CurrentModelKey, int(ollama.Llama3Dot1))
-		loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-			"Creating grocery list with model: "+ollama.ModelName(model).String())
+		loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+			"Creating grocery list")
 		loadingScreen.Show()
 
-		recipe = addMarkdownFomatingToRecipe(recipe)
+		recipe = addMarkdownFormattingToRecipe(recipe)
 
-		generated, err := ollama.AskAI(ollamaClient, ollama.ModelName(model), recipe)
+		generated, err := ollama.AskAI(guiApp, ollamaClient, recipe)
 		if err != nil {
 			slog.Error("Failed to ask AI", "error", err)
 			loadingScreen.Hide()
 			return
 		}
 		loadingScreen.Hide()
-		recipePopUp(guiApp, recipe, &generated)
+		recipePopUp(guiApp, ollamaClient, recipe, &generated)
 	})
 	groceryList.Importance = widget.SuccessImportance
 	budgetFriendlyGroceryList := widget.NewButton("Create a Budget Friendly Grocery List", func() {
 		recipe := createBudgetFriendlyGroceryListPrompt(mealInfo)
 		slog.Info("Budget Friendly Shopping plan", "PROMPT", recipe)
-		model := guiApp.Preferences().IntWithFallback(CurrentModelKey, int(ollama.Llama3Dot1))
-		loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-			"Creating grocery list with model: "+ollama.ModelName(model).String())
+		loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+			"Creating budget friendly grocery list")
 		loadingScreen.Show()
 
-		recipe = addMarkdownFomatingToRecipe(recipe)
+		recipe = addMarkdownFormattingToRecipe(recipe)
 
-		generated, err := ollama.AskAI(ollamaClient, ollama.ModelName(model), recipe)
+		generated, err := ollama.AskAI(guiApp, ollamaClient, recipe)
 		if err != nil {
 			slog.Error("Failed to ask AI", "error", err)
 			loadingScreen.Hide()
 			return
 		}
 		loadingScreen.Hide()
-		recipePopUp(guiApp, recipe, &generated)
+		recipePopUp(guiApp, ollamaClient, recipe, &generated)
 	})
 	budgetFriendlyGroceryList.Importance = widget.SuccessImportance
 
@@ -273,12 +274,12 @@ func shuffleStringArray(str []string) []string {
 	return str
 }
 
-func addMarkdownFomatingToRecipe(recipe string) string {
+func addMarkdownFormattingToRecipe(recipe string) string {
 	return recipe + ". " + "format: markdown"
 }
 
-func recipePopUp(a fyne.App, recipe string, response *ollamaApi.GenerateResponse) {
-	w := a.NewWindow("Ctrl+Revise AI Recipes")
+func recipePopUp(guiApp fyne.App, ollamaClient *ollamaApi.Client, recipe string, response *ollamaApi.GenerateResponse) {
+	w := guiApp.NewWindow("Ctrl+Revise AI Recipes")
 	w.Resize(fyne.NewSize(640, 500))
 
 	hello := widget.NewLabel("Let's Get Cooking!")
@@ -290,83 +291,81 @@ func recipePopUp(a fyne.App, recipe string, response *ollamaApi.GenerateResponse
 
 	vbox := container.NewVScroll(generatedText1)
 
-	model := guiApp.Preferences().IntWithFallback(CurrentModelKey, int(ollama.Llama3Dot1))
-
 	buttons := container.NewPadded(
 		widget.NewButton("Something Different", func() {
-			loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-				"Using model: "+ollama.ModelName(model).String()+" with prompt: "+selectedPrompt.String()+"...")
+			loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+				"Something Different...")
 			loadingScreen.Show()
-			reGenerated, err := ollama.AskAiWithStringAndContext(ollamaClient, ollama.ModelName(model), response.Context,
+			reGenerated, err := ollama.AskAiWithStringAndContext(guiApp, ollamaClient, response.Context,
 				"That doesn't sound good, how about something else please.")
 			if err != nil {
 				slog.Error("Failed to re-generate", "error", err)
 				return
 			}
-			lastClipboardContent = sha256.Sum256([]byte(response.Response))
+			shortcuts.LastClipboardContent = sha256.Sum256([]byte(response.Response))
 			w.Hide()
 			loadingScreen.Hide()
-			recipePopUp(a, recipe, &reGenerated)
+			recipePopUp(guiApp, ollamaClient, recipe, &reGenerated)
 		}),
 		widget.NewButton("Pick a different Protein", func() {
-			loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-				"Using model: "+ollama.ModelName(model).String()+" with prompt: "+selectedPrompt.String()+"...")
+			loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+				"Pick a different Protein...")
 			loadingScreen.Show()
-			reGenerated, err := ollama.AskAiWithStringAndContext(ollamaClient, ollama.ModelName(model), response.Context,
+			reGenerated, err := ollama.AskAiWithStringAndContext(guiApp, ollamaClient, response.Context,
 				"That doesn't sound good, how about something with a different protein please.")
 			if err != nil {
 				slog.Error("Failed to re-generate", "error", err)
 				return
 			}
-			lastClipboardContent = sha256.Sum256([]byte(response.Response))
+			shortcuts.LastClipboardContent = sha256.Sum256([]byte(response.Response))
 			w.Hide()
 			loadingScreen.Hide()
-			recipePopUp(a, recipe, &reGenerated)
+			recipePopUp(guiApp, ollamaClient, recipe, &reGenerated)
 		}),
 		widget.NewButton("Something Simple and Quick", func() {
-			loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-				"Using model: "+ollama.ModelName(model).String()+" with prompt: "+selectedPrompt.String()+"...")
+			loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+				"Something Simple and Quick...")
 			loadingScreen.Show()
-			reGenerated, err := ollama.AskAiWithStringAndContext(ollamaClient, ollama.ModelName(model), response.Context,
+			reGenerated, err := ollama.AskAiWithStringAndContext(guiApp, ollamaClient, response.Context,
 				"That doesn't sound good, how about something that's simple and quick to put together please.")
 			if err != nil {
 				slog.Error("Failed to re-generate", "error", err)
 				return
 			}
-			lastClipboardContent = sha256.Sum256([]byte(response.Response))
+			shortcuts.LastClipboardContent = sha256.Sum256([]byte(response.Response))
 			w.Hide()
 			loadingScreen.Hide()
-			recipePopUp(a, recipe, &reGenerated)
+			recipePopUp(guiApp, ollamaClient, recipe, &reGenerated)
 		}),
 		widget.NewButton("Something Healthy", func() {
-			loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-				"Using model: "+ollama.ModelName(model).String()+" with prompt: "+selectedPrompt.String()+"...")
+			loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+				"Something Healthy...")
 			loadingScreen.Show()
-			reGenerated, err := ollama.AskAiWithStringAndContext(ollamaClient, ollama.ModelName(model), response.Context,
+			reGenerated, err := ollama.AskAiWithStringAndContext(guiApp, ollamaClient, response.Context,
 				"That doesn't sound good, how about something really healthy but still tasty please.")
 			if err != nil {
 				slog.Error("Failed to re-generate", "error", err)
 				return
 			}
-			lastClipboardContent = sha256.Sum256([]byte(response.Response))
+			shortcuts.LastClipboardContent = sha256.Sum256([]byte(response.Response))
 			w.Hide()
 			loadingScreen.Hide()
-			recipePopUp(a, recipe, &reGenerated)
+			recipePopUp(guiApp, ollamaClient, recipe, &reGenerated)
 		}),
 		widget.NewButton("Let's Not Cook", func() {
-			loadingScreen := gui.LoadingScreenWithMessage(guiApp, thinkingMsg,
-				"Using model: "+ollama.ModelName(model).String()+" with prompt: "+selectedPrompt.String()+"...")
+			loadingScreen := loading.LoadingScreenWithMessageAddModel(guiApp, loading.ThinkingMsg,
+				"Let's Not Cook")
 			loadingScreen.Show()
-			reGenerated, err := ollama.AskAiWithStringAndContext(ollamaClient, ollama.ModelName(model), response.Context,
+			reGenerated, err := ollama.AskAiWithStringAndContext(guiApp, ollamaClient, response.Context,
 				"I don't feel like cooking or heating anything up, how about something cold I could put together please.")
 			if err != nil {
 				slog.Error("Failed to re-generate", "error", err)
 				return
 			}
-			lastClipboardContent = sha256.Sum256([]byte(response.Response))
+			shortcuts.LastClipboardContent = sha256.Sum256([]byte(response.Response))
 			w.Hide()
 			loadingScreen.Hide()
-			recipePopUp(a, recipe, &reGenerated)
+			recipePopUp(guiApp, ollamaClient, recipe, &reGenerated)
 		}),
 		widget.NewButtonWithIcon("Copy the Recipe to Clipboard", theme.ContentCopyIcon(), func() {
 			w.Clipboard().SetContent(response.Response)
@@ -386,4 +385,26 @@ func recipePopUp(a fyne.App, recipe string, response *ollamaApi.GenerateResponse
 		container.NewVScroll(grid),
 	))
 	w.Show()
+}
+
+func footer() *fyne.Container {
+	footer := container.NewHBox(
+		layout.NewSpacer(),
+		widget.NewHyperlink("Ctrl+Revise", parseURL("https://ctrlplusrevise.com")),
+		widget.NewLabel("-"),
+		widget.NewHyperlink("Documentation", parseURL("https://ctrlplusrevise.com/docs/tutorials/")),
+		widget.NewLabel("-"),
+		widget.NewHyperlink("Sponsor", parseURL("https://www.patreon.com/SalmonsStudios")),
+		layout.NewSpacer(),
+	)
+	return footer
+}
+
+func parseURL(urlStr string) *url.URL {
+	link, err := url.Parse(urlStr)
+	if err != nil {
+		fyne.LogError("Could not parse URL", err)
+	}
+
+	return link
 }
