+++ internal/gui/food/meal_planner_prompts.go
@@ -1,4 +1,4 @@
-package main
+package food
 
 import (
 	"fmt"
