+++ internal/gui/loading/loading_screen.go
@@ -1,10 +1,17 @@
-package gui
+package loading
 
 import (
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/container"
 	"fyne.io/fyne/v2/data/binding"
 	"fyne.io/fyne/v2/widget"
+
+	"github.com/bahelit/ctrl_plus_revise/internal/config"
+	"github.com/bahelit/ctrl_plus_revise/internal/ollama"
+)
+
+const (
+	ThinkingMsg = "Thinking..."
 )
 
 func ShowNotification(guiApp fyne.App, title, content string) {
@@ -29,6 +36,18 @@ func LoadingScreenWithMessage(guiApp fyne.App, title, msg string) fyne.Window {
 	loadingScreen.SetContent(container.NewVBox(text, infinite))
 	return loadingScreen
 }
+
+func LoadingScreenWithMessageAddModel(guiApp fyne.App, title, msg string) fyne.Window {
+	model := guiApp.Preferences().IntWithFallback(config.CurrentModelKey, int(ollama.Llama3Dot1))
+	modelMsg := "\nUsing model: " + ollama.ModelName(model).String() + "..."
+	title += modelMsg
+	loadingScreen := guiApp.NewWindow(title)
+	infinite := widget.NewProgressBarInfinite()
+	text := widget.NewLabel(msg)
+	loadingScreen.SetContent(container.NewVBox(text, infinite))
+	return loadingScreen
+}
+
 func LoadingScreenWithProgressAndMessage(guiApp fyne.App, loading *widget.ProgressBar, status binding.String, title, msg string) fyne.Window {
 	loadingScreen := guiApp.NewWindow(title)
 	loadingScreen.Resize(fyne.NewSize(300, 80))
